requirements.txt
# ─── Dependencias Principales / Main Dependencies ───────────────────
ultralytics==8.4.23
streamlit>=1.30.0
Pillow>=10.0.0

# ─── Nota sobre Aceleración GPU / GPU Acceleration Note ──────────────
#
# [ES] Las dependencias anteriores instalan PyTorch en modo CPU por defecto.
#      Si tienes una GPU NVIDIA compatible con CUDA, desinstala la versión
#      estándar de PyTorch y ejecuta el siguiente comando:
#
# [EN] The dependencies above install CPU-only PyTorch by default.
#      If you have a CUDA-compatible NVIDIA GPU, uninstall the default
#      PyTorch version and run the following command:
#
# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu130
